api/requirements.txt
aiofile==3.9.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.87.0
anyio==4.13.0
attrs==26.1.0
Authlib==1.6.9
beartype==0.22.9
cachetools==7.0.5
caio==0.9.25
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
cyclopts==4.10.1
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
docutils==0.22.4
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.2
fastmcp==3.2.0
fastuuid==0.14.0
filelock==3.25.2
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.3.0
geographiclib==2.1
geopy==2.4.1
google-auth==2.49.1
google-genai==1.70.0
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.8.0
idna==3.11
importlib_metadata==8.7.1
iniconfig==2.3.0
jaraco.classes==3.4.0
jaraco.context==6.1.2
jaraco.functools==4.4.0
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-path==0.4.5
jsonschema-specifications==2025.9.1
keyring==25.7.0
langchain==1.2.14
langchain-anthropic==1.4.0
langchain-core==1.2.23
langchain-google-genai==4.2.1
langchain-mcp-adapters==0.2.2
langchain-openai==1.1.12
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.23
litellm==1.82.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
more-itertools==10.8.0
multidict==6.7.1
openai==2.30.0
openapi-pydantic==0.5.1
opentelemetry-api==1.40.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pathable==0.5.0
platformdirs==4.9.4
pluggy==1.6.0
propcache==0.4.1
py-key-value-aio==0.4.4
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
pyperclip==1.11.0
pytest==9.0.2
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.3.32
requests==2.33.0
requests-toolbelt==1.0.0
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==3.3.4
starlette==1.0.0
tenacity==9.1.4
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uncalled-for==0.2.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
